Added AppDataReducer
